backend/requirements.txt
fastapi
sqlmodel
psycopg2-binary
python-dotenv
uvicorn[standard]
bcrypt
email-validator
python-jose[cryptography]
python-multipart
pandas
pinecone
openai
PyPDF2
python-docx
openpyxl
websockets
langgraph
langchain
langchain-core
langchain-openai
langchain-pinecone
alembic
beautifulsoup4
requests
html5lib
lxml
celery[redis]
redis